requirements.txt
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
duplicity==0.8.21
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.0
fastapi-cli==0.0.5
fasteners==0.14.1
fastjsonschema==2.20.0
fqdn==1.5.1
future==0.18.2
groq==0.11.0
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.2.2
notebook_shim==0.2.4
oauthlib==3.2.0
olefile==0.46
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.47
protobuf==3.12.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.10
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
SecretStorage==3.3.1
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.6
systemd-python==234
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.22.0
uvloop==0.20.0
wadllib==1.3.6
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
xdg==5
xkit==0.0.0
zipp==1.0.0
groq
httpx
fastapi
pydantic
flask
redis
